extract fundBounty method in separate file
